+++ src/refactoring/hooks/cart/useCart.ts
@@ -1,6 +1,6 @@
 import { useState } from "react";
-import { CartItem, Coupon, Product } from "../../types";
-import { calculateCartTotal, updateCartItemQuantity } from "./utils/cartUtils";
+import { CartItem, Coupon, Product } from "../../../types";
+import { calculateCartTotal, updateCartItemQuantity } from "../utils/cartUtils";
 
 export const useCart = () => {
   const [cart, setCart] = useState<CartItem[]>([]);
