+++ src/refactoring/hooks/coupon/useCoupon.ts
@@ -1,4 +1,4 @@
-import { Coupon } from "../../types.ts";
+import { Coupon } from "../../../types.ts";
 import { useState } from "react";
 
 export const useCoupons = (initialCoupons: Coupon[]) => {
