+++ src/refactoring/hooks/coupon/useCouponManagement.ts
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { Coupon } from "../../types";
+import { Coupon } from "../../../types";
 
 export const useCouponManagement = (
   initialCoupons: Coupon[],
