+++ src/refactoring/hooks/product/useProduct.ts
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { Product } from "../../types.ts";
+import { Product } from "../../../types.ts";
 
 export const useProducts = (initialProducts: Product[]) => {
   const [products, setProducts] = useState<Product[]>(initialProducts);
