+++ src/refactoring/hooks/product/useProductAccordion.ts
